DOC:  Applied suggestions from the review
